fix: next-font not working
