+++ src/pages/_app.tsx
@@ -1,6 +1,7 @@
 import '@code-hike/mdx/dist/index.css';
 import 'mod/styles/globals.css';
 
+import localFont from '@next/font/local';
 import { Analytics } from '@vercel/analytics/react';
 import { getAnalytics } from 'firebase/analytics';
 import { getApp, getApps, initializeApp } from 'firebase/app';
@@ -12,16 +13,26 @@ import type { FC } from 'react';
 import Layout from 'mod/components/Layout';
 import defaultSeoConfig from 'mod/lib/next-seo.config';
 
-const app =
-  getApps().length <= 0
-    ? initializeApp({ ...JSON.parse(process.env.NEXT_PUBLIC_FIREBASE_CRED) })
-    : getApp();
-
 const Head: FC = () => (
   <>
     <DefaultSeo {...defaultSeoConfig} />
   </>
 );
+
+const switzer = localFont({
+  src: [
+    { path: '../assets/switzer.woff2', style: 'normal' },
+    { path: '../assets/switzer-italic.woff2', style: 'italic' }
+  ],
+  display: 'swap',
+  preload: true,
+  variable: '--font-switzer'
+});
+
+const app =
+  getApps().length <= 0
+    ? initializeApp({ ...JSON.parse(process.env.NEXT_PUBLIC_FIREBASE_CRED) })
+    : getApp();
 
 export default function App({ Component, pageProps }: AppProps) {
   if (typeof window !== 'undefined' && process.env.NODE_ENV === 'production') {
@@ -32,7 +43,7 @@ export default function App({ Component, pageProps }: AppProps) {
   return (
     <>
       <Head />
-      <Layout>
+      <Layout fontFamily={switzer}>
         <Component {...pageProps} />
       </Layout>
       <Analytics />
